move affine.apply lowering to rewrite utils
