+++ lib/Utils/RewriteUtils/BUILD
@@ -9,6 +9,8 @@ cc_library(
     hdrs = ["RewriteUtils.h"],
     deps = [
         "@llvm-project//llvm:Support",
+        "@llvm-project//mlir:AffineDialect",
+        "@llvm-project//mlir:AffineUtils",
         "@llvm-project//mlir:IR",
     ],
 )
